+++ vscode-dotnet-runtime-extension/src/extension.ts
@@ -43,6 +43,18 @@ import {
     GlobalInstallerResolver,
     SdkInstallationDirectoryProvider,
     CommandExecutor,
+    IDotnetListVersionsContext,
+    WebRequestWorker,
+    IDotnetVersion,
+    DotnetVersionResolutionError,
+    IDotnetListVersionsResult,
+    LinuxVersionResolver,
+    LinuxInstallType,
+    GlobalAcquisitionContextMenuOpened,
+    UserManualInstallVersionChosen,
+    UserManualInstallRequested,
+    UserManualInstallSuccess,
+    UserManualInstallFailure,
 } from 'vscode-dotnet-runtime-library';
 import { dotnetCoreAcquisitionExtensionId } from './DotnetCoreAcquisitionId';
 import { IAcquisitionWorkerContext } from 'vscode-dotnet-runtime-library/dist/Acquisition/IAcquisitionWorkerContext';
@@ -62,6 +74,9 @@ namespace commandKeys {
     export const acquireGlobalSDK = 'acquireGlobalSDK';
     export const acquireStatus = 'acquireStatus';
     export const uninstallAll = 'uninstallAll';
+    export const listVersions = 'listVersions';
+    export const recommendedVersion = 'recommendedVersion'
+    export const globalAcquireSDKPublic = 'acquireGlobalSDKPublic';
     export const showAcquisitionLog = 'showAcquisitionLog';
     export const ensureDotnetDependencies = 'ensureDotnetDependencies';
     export const reportIssue = 'reportIssue';
@@ -120,6 +135,8 @@ export function activate(context: vscode.ExtensionContext, extensionContext?: IE
     const sdkContext = getAcquisitionWorkerContext(false);
     const sdkIssueContextFunctor = getIssueContext(existingPathConfigWorker);
     const sdkAcquisitionWorker = getAcquisitionWorker(sdkContext);
+
+    checkIfSDKAcquisitionIsSupported();
 
     // Creating API Surfaces
     const dotnetAcquireRegistration = vscode.commands.registerCommand(`${commandPrefix}.${commandKeys.acquire}`, async (commandContext: IDotnetAcquireContext) => {
@@ -194,6 +211,65 @@ export function activate(context: vscode.ExtensionContext, extensionContext?: IE
         return pathResult;
     });
 
+    const dotnetListVersionsRegistration = vscode.commands.registerCommand(`${commandPrefix}.${commandKeys.listVersions}`,
+    async (commandContext: IDotnetListVersionsContext | undefined, customWebWorker: WebRequestWorker | undefined) =>
+    {
+        return getAvailableVersions(commandContext, customWebWorker, false);
+    });
+
+    const dotnetRecommendedVersionRegistration = vscode.commands.registerCommand(`${commandPrefix}.${commandKeys.recommendedVersion}`,
+    async (commandContext: IDotnetListVersionsContext | undefined, customWebWorker: WebRequestWorker | undefined) : Promise<string> =>
+    {
+        const availableVersions = await getAvailableVersions(commandContext, customWebWorker, true);
+        const activeSupportVersions = availableVersions?.filter( (version : IDotnetVersion) => version.supportPhase === 'active');
+
+        if (!activeSupportVersions || activeSupportVersions.length < 1)
+        {
+            const err = new Error(`An active-support version of dotnet couldn't be found. Discovered versions: ${JSON.stringify(availableVersions)}`);
+            globalEventStream.post(new DotnetVersionResolutionError(err as Error, 'recommended'));
+            if(!availableVersions || availableVersions.length < 1)
+            {
+                return '';
+            }
+            return availableVersions[0].version;
+        }
+
+        // The first item will be the newest version.
+        return activeSupportVersions[0].version;
+    });
+
+    const acquireGlobalSDKPublicRegistration = vscode.commands.registerCommand(`${commandPrefix}.${commandKeys.globalAcquireSDKPublic}`, async (commandContext: IDotnetAcquireContext) =>
+    {
+        globalEventStream.post(new GlobalAcquisitionContextMenuOpened(`The user has opened the global SDK acquisition context menu.`));
+
+        const recommendedVersion : string = await vscode.commands.executeCommand('dotnet.recommendedVersion') ?? '';
+
+        const chosenVersion = await vscode.window.showInputBox(
+        {
+                placeHolder: recommendedVersion,
+                value: recommendedVersion,
+                prompt: 'The .NET SDK version. You can use different formats: 5, 3.1, 7.0.3xx, 6.0.201, etc.',
+        }) ?? '';
+
+        globalEventStream.post(new UserManualInstallVersionChosen(`The user has chosen to install the .NET SDK version ${chosenVersion}.`));
+
+        try
+        {
+            globalEventStream.post(new UserManualInstallRequested(`Starting to install the .NET SDK ${chosenVersion} via a user request.`));
+
+            await vscode.commands.executeCommand('dotnet.showAcquisitionLog');
+            const userCommandContext : IDotnetAcquireContext = { version: chosenVersion, requestingExtensionId: 'user', installType: 'global' };
+            await vscode.commands.executeCommand('dotnet.acquireGlobalSDK', userCommandContext);
+            globalEventStream.post(new UserManualInstallSuccess(`The .NET SDK ${chosenVersion} was successfully installed.`));
+
+        }
+        catch (error)
+        {
+            globalEventStream.post(new UserManualInstallFailure((error as Error), `The .NET SDK ${chosenVersion} failed to install. Error: ${(error as Error).toString()}`));
+            vscode.window.showErrorMessage((error as Error).toString());
+        }
+    });
+
     const dotnetAcquireStatusRegistration = vscode.commands.registerCommand(`${commandPrefix}.${commandKeys.acquireStatus}`, async (commandContext: IDotnetAcquireContext) => {
         const pathResult = callWithErrorHandling(async () => {
             globalEventStream.post(new DotnetAcquisitionStatusRequested(commandContext.version, commandContext.requestingExtensionId));
@@ -249,6 +325,47 @@ export function activate(context: vscode.ExtensionContext, extensionContext?: IE
         });
     }
 
+    function checkIfSDKAcquisitionIsSupported() : boolean
+    {
+        let isSupported = true;
+        isSupported = isSupported && !CommandExecutor.isRunningUnderWSL();
+        vscode.commands.executeCommand('setContext', 'dotnetAcquisitionExtension.isGlobalSDKSupported', isSupported);
+        return isSupported;
+    }
+
+    const getAvailableVersions = async (commandContext: IDotnetListVersionsContext | undefined,
+        customWebWorker: WebRequestWorker | undefined, onRecommendationMode : boolean) : Promise<IDotnetListVersionsResult | undefined> =>
+    {
+        const customVersionResolver = new VersionResolver(sdkContext, customWebWorker);
+
+        if(os.platform() !== 'linux' || !onRecommendationMode)
+        {
+            const versionsResult = await callWithErrorHandling(async () =>
+            {
+                return customVersionResolver.GetAvailableDotnetVersions(commandContext);
+            }, getIssueContext(existingPathConfigWorker)(commandContext?.errorConfiguration, 'getAvailableVersions'));
+
+            return versionsResult;
+        }
+        else
+        {
+            const linuxResolver = new LinuxVersionResolver(sdkContext, utilContext);
+            const suggestedVersion = await linuxResolver.getRecommendedDotnetVersion('sdk' as LinuxInstallType);
+            const osAgnosticVersionData = await getAvailableVersions(commandContext, customWebWorker, !onRecommendationMode);
+            const resolvedSupportPhase = osAgnosticVersionData?.find((version : IDotnetVersion) =>
+                customVersionResolver.getMajorMinor(version.version) === customVersionResolver.getMajorMinor(suggestedVersion))?.supportPhase ?? 'active';
+                // Assumption : The newest version is 'active' support, but we can't gaurantee that.
+                // If the linux version is too old it will eventually support no active versions of .NET, which would cause a failure.
+                // The best we can give it is the newest working version, which is the most likely to be supported, and mark it as active so we can use it.
+
+            return [
+                { version: suggestedVersion, channelVersion: `${customVersionResolver.getMajorMinor(suggestedVersion)}`,
+                supportStatus: Number(customVersionResolver.getMajor(suggestedVersion)) % 2 === 0 ? 'lts' : 'sts',
+                supportPhase: resolvedSupportPhase }
+            ];
+        }
+    }
+
     function getAcquisitionWorkerContext(isRuntimeInstall : boolean) : IAcquisitionWorkerContext
     {
         return {
@@ -290,6 +407,9 @@ export function activate(context: vscode.ExtensionContext, extensionContext?: IE
         dotnetAcquireRegistration,
         dotnetAcquireStatusRegistration,
         dotnetAcquireGlobalSDKRegistration,
+        acquireGlobalSDKPublicRegistration,
+        dotnetListVersionsRegistration,
+        dotnetRecommendedVersionRegistration,
         dotnetUninstallAllRegistration,
         showOutputChannelRegistration,
         ensureDependenciesRegistration,